Move App into the components folder & replace the root div with a react fragment.
